+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.3.1
 ------
 - Lightning encoding for RGB types used in lightning messages
 - Added missed StrictEncoding derives
+- Consignment ids & bech32 representation
 
 v0.3.0
 ------